requirements.txt
lxml==5.2.2
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-docx==1.1.2
typing_extensions==4.12.2
